Update date.py
Found another issue, this time in to_java_calendar().  When passing in some date time types (usually python) the timezone/offset information was lost, which would result in the calendar being returned without any offset, it would just assume it was 'GMT'.  The java.util.TimeZone.getTimeZone() method is looking specifically for either 'America/Chicago' type zoneId, or 'GMT-05:00'.  'GMT' is missing from some date time type zoneId's.  I added a regex that will search for the format '(+/-)[00-27]:[00-59]' If it is a match, it will add 'GMT' to the front of the zoneId string.