Update date.py
Update to elapsed().  I didn't realize that function calls that are made in default arguments within functions are called at script loading and not at the function's run time.  So calling elapsed(SomeTime, end=joda_now()) would actually give an end time when the whole script was originally loaded.  I moved it within the function and test for end == None.