when: use scriptExtension instead of scope
If a rule is created at runtime and the function resides in a file in
`lib/python`, when that rule is triggered there will be no scope
available. If `when` is used by that rule to generate another rule, it
will not find needed imports in `jsr223.scope`. If `core.triggers` was
imported at any time by the script that generated the original rule,
`core.triggers` will have an instance of `scriptExtension` from that
original scope that can be used to get things from that scope instead.